Refactoring TimeSeries objects
Moving TimeSeries structs and enums from asap.rs to crates/time-series.
Modifying lttb.rs to use TimeSeries.
